requirements.txt
langchain>=0.1.0
langchain-groq>=1.1.0
langchain-core>=0.1.10
python-dotenv>=1.0.0
